pydantic-ai/pydantic-ai-examples/pyproject.toml
[project]
name = "pydantic-ai-examples"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "martimfasantos", email = "72747170+martimfasantos@users.noreply.github.com"},
]
dependencies = ["pydantic-ai-slim[logfire,openai,vertexai]>=0.0.24", "pydantic-settings>=2.7.1"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.logfire]
ignore_no_config = true

[tool.pdm.scripts]
tools = "python3 src/pydantic_ai_examples/1_tools_and_dependency_injection.py"
parallel = "python3 src/pydantic_ai_examples/2_parallel_tool_calling.py"
programmatic-handoff = "python3 src/pydantic_ai_examples/3_programmatic_handoff.py"
graph-gen-ai = "python3 src/pydantic_ai_examples/4_graph_gen_ai.py"

